[master] Reorganize code about DubboConfigEarlyRegistrationPostProcessor
